Move auth classes into Northstar\Auth namespace.
